requirement.txt
ChatterBot-0.8.7
simple-websocket-server
